requirements.txt
# Web 框架
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# HTTP 请求
requests>=2.31.0
aiohttp>=3.9.0  # 异步 HTTP 请求

# 网页解析
beautifulsoup4>=4.12.0

# 浏览器自动化（登录时需要）
selenium>=4.15.0

# 配置文件
pyyaml>=6.0.0

# 异步支持
anyio>=4.0.0

